derp

+++ uninstall.sh
@@ -1,5 +1,9 @@
 echo 'Uninstalling Twister OS Updater by FlameKat53 and MobileGamesMotionYT...'
 echo 'Thanks for using me! You can install me again at "https://bit.ly/twistpatch"'
 cd ~/
+if [ -d ~/updater/ ]; then
 rm -r ~/updater/
+fi
+if [ -f ~/.local/share/applications/updater.desktop ]; then
 rm ~/.local/share/applications/updater.desktop
+fi
